do not remove stale repo permissions or org memberships if empty

+++ lib/travis/github/sync/organizations.rb
@@ -6,7 +6,7 @@ module Travis
       class Organizations
         class << self
           def cancel_memberships(user, orgs)
-            user.memberships.where(:organization_id => orgs.map(&:id)).delete_all
+            user.memberships.where(:organization_id => orgs.map(&:id)).delete_all unless orgs.empty?
           end
         end
